pyproject.toml
[project]
name = "dl-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.3.0",
    "mcp>=1.20.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
